+++ src/asp/Sessions/StereoSessionNadirPinhole.h
@@ -45,7 +45,9 @@ namespace asp {
 
     static StereoSession* construct() { return new StereoSessionNadirPinhole; }
 
-    virtual bool have_datum() const {return !asp::stereo_settings().no_datum; }
+    virtual bool have_datum() const {
+      return !asp::stereo_settings().no_datum && !stereo_settings().correlator_mode;
+    }
 
     // Unlike non-nadir pinhole, this follows the default method.
     virtual tx_type tx_left () const {return tx_left_homography ();}